Reduce hardlockup probability when sending dm updates
Prior to these changes, the updates on the data model
were send to the remote site while holding the slc lock
and with irqs disabled.
Under heavy load this will likely lead to a hardlockup,
because the sender busy waits until a slot in the ringbuffer becomes
available.
However, there is no need to hold the slc lock while trying
to push the dm update. Just the ringbuffer lock is needed.
Therefore, the dm update is now sent after performing the actual
local dm update, and without the slc lock beeing hold.
Moreover, the locking tim in *registerProvider function was reduced as well.